feat(task-044): extract transformer, diagnostics, and sandbox helpers from analysis_executor.py
- Move ProactiveCodeTransformer to analysis_transformer.py (stateless, zero-arg)
- Move prerequisite checks, error recovery, LLM diagnosis, output validation,
  and package management to analysis_diagnostics.py
- Move sandbox globals, execution entry points, approval gating, doublet
  harmonization, figure autosave, and context directives to analysis_sandbox.py
- Rewrite analysis_executor.py as thin facade (~150 lines) delegating to helpers
- Re-export ProactiveCodeTransformer and _PACKAGE_ALIASES from facade for compat
- Add test_analysis_decomposition.py with 31 focused decomposition tests
- Update test_decomposition_contracts.py known-modules list

Acceptance criteria: analysis_executor.py no longer carries full implementations
inline; AnalysisExecutor remains the stable import facade; all 423 tests pass.
